requirements.txt
streamlit
duckduckgo_search
requests
Pillow